Move actions to separate packages
