+++ android/libraries/rib-base/src/main/java/com/badoo/ribs/core/routing/backstack/ConfigurationCommand.kt
@@ -3,15 +3,15 @@ package com.badoo.ribs.core.routing.backstack
 import android.os.Parcelable
 import com.badoo.ribs.core.Node
 import com.badoo.ribs.core.routing.action.RoutingAction
-import com.badoo.ribs.core.routing.backstack.action.ActivateAction
-import com.badoo.ribs.core.routing.backstack.action.AddAction
-import com.badoo.ribs.core.routing.backstack.action.DeactivateAction
-import com.badoo.ribs.core.routing.backstack.action.MultiConfigurationAction
-import com.badoo.ribs.core.routing.backstack.action.NoOpAction
-import com.badoo.ribs.core.routing.backstack.action.RemoveAction
-import com.badoo.ribs.core.routing.backstack.action.SingleConfigurationAction
-import com.badoo.ribs.core.routing.backstack.action.SleepAction
-import com.badoo.ribs.core.routing.backstack.action.WakeUpAction
+import com.badoo.ribs.core.routing.backstack.action.single.ActivateAction
+import com.badoo.ribs.core.routing.backstack.action.single.AddAction
+import com.badoo.ribs.core.routing.backstack.action.single.DeactivateAction
+import com.badoo.ribs.core.routing.backstack.action.multi.MultiConfigurationAction
+import com.badoo.ribs.core.routing.backstack.action.single.NoOpAction
+import com.badoo.ribs.core.routing.backstack.action.single.RemoveAction
+import com.badoo.ribs.core.routing.backstack.action.single.SingleConfigurationAction
+import com.badoo.ribs.core.routing.backstack.action.multi.SleepAction
+import com.badoo.ribs.core.routing.backstack.action.multi.WakeUpAction
 
 /**
  * Represents a command to change one or more [ConfigurationContext] elements.
@@ -30,11 +30,13 @@ internal sealed class ConfigurationCommand<C : Parcelable> {
         abstract val action: MultiConfigurationAction<C>
 
         class Sleep<C : Parcelable> : MultiConfigurationCommand<C>() {
-            override val action: MultiConfigurationAction<C> = SleepAction()
+            override val action: MultiConfigurationAction<C> =
+                SleepAction()
         }
 
         class WakeUp<C : Parcelable> : MultiConfigurationCommand<C>() {
-            override val action: MultiConfigurationAction<C> = WakeUpAction()
+            override val action: MultiConfigurationAction<C> =
+                WakeUpAction()
         }
     }
 
@@ -47,19 +49,23 @@ internal sealed class ConfigurationCommand<C : Parcelable> {
              * No additional action here, as [AddAction] is executed automatically
              * during [ConfigurationContext.Unresolved.resolve],
              */
-            override val action: SingleConfigurationAction = NoOpAction
+            override val action: SingleConfigurationAction =
+                NoOpAction
         }
 
         data class Activate<C : Parcelable>(override val key: ConfigurationKey) : SingleConfigurationCommand<C>() {
-            override val action: SingleConfigurationAction = ActivateAction
+            override val action: SingleConfigurationAction =
+                ActivateAction
         }
 
         data class Deactivate<C : Parcelable>(override val key: ConfigurationKey) : SingleConfigurationCommand<C>() {
-            override val action: SingleConfigurationAction = DeactivateAction
+            override val action: SingleConfigurationAction =
+                DeactivateAction
         }
 
         data class Remove<C : Parcelable>(override val key: ConfigurationKey) : SingleConfigurationCommand<C>() {
-            override val action: SingleConfigurationAction = RemoveAction
+            override val action: SingleConfigurationAction =
+                RemoveAction
         }
     }
 }
