+++ android/libraries/rib-base/src/main/java/com/badoo/ribs/core/routing/backstack/action/single/ActivateAction.kt
@@ -1,10 +1,11 @@
-package com.badoo.ribs.core.routing.backstack.action
+package com.badoo.ribs.core.routing.backstack.action.single
 
 import android.os.Parcelable
 import com.badoo.ribs.core.Node
 import com.badoo.ribs.core.routing.action.RoutingAction
 import com.badoo.ribs.core.routing.backstack.ConfigurationContext
 import com.badoo.ribs.core.routing.backstack.ConfigurationContext.Resolved
+import com.badoo.ribs.core.routing.backstack.action.ActionExecutionParams
 import com.badoo.ribs.core.routing.backstack.feature.WorkingState
 
 /**
