+++ android/libraries/rib-base/src/main/java/com/badoo/ribs/core/routing/backstack/action/single/RemoveAction.kt
@@ -1,8 +1,9 @@
-package com.badoo.ribs.core.routing.backstack.action
+package com.badoo.ribs.core.routing.backstack.action.single
 
 import android.os.Parcelable
 import com.badoo.ribs.core.Node
 import com.badoo.ribs.core.routing.backstack.ConfigurationContext.Resolved
+import com.badoo.ribs.core.routing.backstack.action.ActionExecutionParams
 
 /**
  * Removes [Node]s from their parent, resulting in the end of their lifecycles.
